Changed GL_ARB_texture_float to GL_ARB_texture_rg. Now compiles again in OSX
git-svn-id: https://irrlicht.svn.sourceforge.net/svnroot/irrlicht/branches/releases/1.6@2632 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/COpenGLShaderMaterialRenderer.cpp
@@ -206,7 +206,7 @@ bool COpenGLShaderMaterialRenderer::checkError(const irr::c8* type)
 
 		core::stringc errString = type;
 		errString += " compilation failed at position ";
-		errString += core::stringc(errPos);
+		errString += core::stringc(s32(errPos));
 		errString += ":\n";
 		errString += reinterpret_cast<const char*>(glGetString(GL_PROGRAM_ERROR_STRING_ARB));
